Add files via upload
Business Objective: To understand the user behaviour and engagement on Trell, a social media platform that focuses on short-form video content, and develop a machine learning model that can predict user age group and use it for target marketing.
Approach: The dataset of Trell social media is taken from Cascade Cup Data Science Hackathon. Data set consists of the user details and their activities in the app. ‘Stratified sampling’ was used to maintain the balance of the target variable. Various classification and ensemble learning models has been built on the data which was evaluated using the f1 score and attained a final F1 score of 0.75.
Tools and Libraries Used: Python, NumPy, Pandas, Matplotlib, Seaborn, Sklearn